Bugfix: Z80 Monitor: trace / go start address
When starting a trace ("t") or go ("g"), do not start at
the currently displayed address, since that address changes
when using the unassemble ("u") or data display ("d") command.
Instead, use the current contents of the PC register (program
counter) as start address, unless a start address is
explicitly specified together with the trace or go command.