Feature: Add mouse listener for keyboard panel keys
Some keys may not not be natively available over a
standard PC keyboard layout, e.g. when the key is
accessible with shift key only on the PC keyboard
layout, while on the VZ200 keyboard layout, it is a
base key to enter without pressing shift.

For these cases, it is vital to provide alternative
access.  This commit adds a mouse listener for the
VZ200 keyboard panel buttons.  This way, you may
also access a VZ200's shift key by holding the
shift key on your PC keyboard and clicking on the
VZ200's keyboard panel into the target key.