Bugfix: Get rid of separate timer thread
Rather than running I/O related timing in a
separate thread, perform it synchronously with
execution of CPU instructions to get more exact
timing.  Now, time base is the number of CPU
cycles that have been executed since start of
the emulator software.  I/O may now sync its
timing to this time base.

As a side effect, when stopping execution in the
monitor, generation of I/O interrupts will also
stop, such that when resuming execution, the next
instruction will not always be a call to an
interrupt service routine, but flow control will
continue ordinarily, as if execution had not been
stopped before.