Z80: Bugfix: Partial "DAA" implementation
The "DAA" instruction is poorly documented
in Z80 specifications available for me, such
that it is extremely difficult to correctly
implement it.  On the other hand, this
instruction is very seldom used.
For this reason, I currently implement this
instruction in the 8080 flavour, thereby
intentionally ignoring the processor status
"N" flag, which must be considered for a
correct Z80-style implementation of the
"DAA" instruction.
If better documentation is available for
the Z80-style version of the "DAA"
instruction, I will be happy to correct
the current implementation.